% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar_power.R
\name{mcnemar_power}
\alias{mcnemar_power}
\title{Power Simulations for McNemar's Test for Sensitivity/Specificity}
\usage{
mcnemar_power(
  N = 1000,
  n.sims = 1000,
  alpha = 0.05,
  seed = 12345,
  prev = p,
  sens1 = sn1,
  spec1 = sp1,
  sens2 = sn2,
  spec2 = sp2
)
}
\arguments{
\item{N}{vector of candidate sample sizes}

\item{n.sims}{number of simulations to run for each N}

\item{alpha}{significance level for McNemar}

\item{seed}{set yo seed!}

\item{prev}{true positive prevalence}

\item{sens1}{sensitivity of first test}

\item{spec1}{specificity of first test}

\item{sens2}{sensitivity of second test}

\item{spec2}{specificity of second test}
}
\description{
This function takes values for sensitivity and specificity of
two tests for comparison and calculates the power, given n,
needed to observe significant differences in the performance
between the two. Power is returned for both sensitivity and
specificity by running McNemar's test on simulated true positives
and true negatives respectively. Returns a matrix with simulated
power given sample size.
}
