% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosamp_power.R
\name{twosamp_power}
\alias{twosamp_power}
\title{Power Simulations for Two-Sample Comparisons Tests (T-test or Wilcox Rank-Sum)}
\usage{
twosamp_power(
  N = 1000,
  n.sims = 1000,
  alpha = 0.05,
  seed = 12345,
  par = F,
  dist1 = a,
  dist2 = b
)
}
\arguments{
\item{N}{Vector representing sample size per group}

\item{n.sims}{Number of simulations to run for each N}

\item{alpha}{Significance level for Wilcox Test}

\item{seed}{Set your seed}

\item{par}{'T' for parametric T-test, 'F' for non-parametric Wilcox Rank-Sum test}

\item{dist1}{Vector containing values (either real or simulated) for first distribution being compared}

\item{dist2}{Vector containing values (either real or simulated) for second distribution being compared}
}
\description{
This function takes inputs for two distributions of data (as vectors) and runs either
a parametric two-sample T-Test or non-parametric Wilcox rank-sum test. Each simulation
samples from the candidate distributions with replacement for each iteration. Returns
data frame of simulated power for each candidate N. Will add functionality for differing 
sample sizes.
}
